Seperate the Voting page into VotingPage and VotingPageContent
